pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ratscramble"
version = "0.1.0"
description = "AI-agent simulation of Rat Scramble"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "boto3>=1.34.0",
  "pydantic>=2.8.0",
  "rich>=13.7.0",
  "PyYAML>=6.0.1",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.0",
  "pytest-asyncio>=0.23.0",
]

[tool.pytest.ini_options]
pythonpath = [".", "src"]
testpaths = ["tests"]
markers = [
  "bedrock: integration tests that call AWS Bedrock",
]
